backend/requirements.txt
google-generativeai==0.8.3
elevenlabs==1.10.1
pygame==2.6.1
requests==2.32.3
python-dotenv==1.0.1
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
